# PERF: Using Numpy C-API for left-join calls