@jreback 2020-03-19
again I'd rather not do this unless this is a huge perf gain. can you time.