@MomIsBestFriend 2020-03-19
Closing this as there is no performance gain.

I will open a seperate PR addressing comments that were mentioned here